requirements.txt
flask
gunicorn
google.generativeai
numpy
textblob